market-api finished

+++ src/main/java/com/db/marketapi/service/OrderService.java
@@ -2,8 +2,10 @@ package com.db.marketapi.service;
 
 import com.db.marketapi.model.Cart;
 import com.db.marketapi.model.Order;
+import com.db.marketapi.model.User;
 import com.db.marketapi.repository.CartRepository;
 import com.db.marketapi.repository.OrderRepository;
+import com.db.marketapi.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,13 +18,20 @@ import java.util.Optional;
 public class OrderService {
 
   private final OrderRepository orderRepository;
+  private final UserRepository userRepository;
 
   public Order createOrder(Order order) {
     return orderRepository.save(order);
   }
 
-  public void deleteOrderById(Long id) {
-    orderRepository.deleteById(id);
+  public void deleteOrderFromUserOrderHistory(Long userId, Long id) {
+    Optional<User> user = userRepository.findById(userId);
+    Optional<Order> order = orderRepository.findById(id);
+
+    if (user.isPresent()) {
+      user.get().getOrderHistory().remove(order.get());
+      orderRepository.deleteById(id);
+    }
   }
 
   public Order updateOrder(Order order) {
